pyproject.toml
[project]
name = "cibmtr"
version = "0.1.0"
description = "CIBMTR - Equity in post-HCT Survival Predictions"
authors = [{ name = "syamoji141"}]
dependencies = [
    "numpy>=1.24.4",
    "matplotlib>=3.7.2",
    "tqdm>=4.65.0",
    "torch==2.3.0",
    "joblib>=1.3.1",
    "pyyaml>=6.0.1",
    "types-PyYAML>=6.0.12.11",
    "hydra-core>=1.3.2",
    "pytorch_pfn_extras>=0.7.1",
    "scipy>=1.10.1",
    "pandas>=2.0.3",
    #    "pytorch-lightning>=2.0.6",
    "lightning>=2.4.0",
    "polars~=1.14.0",
    "scikit-learn>=1.3.0",
    "streamlit>=1.25.0",
    "kaggle~=1.5.16",
    "jupyter-black~=0.3.4",
    "japanize_matplotlib~=1.1.3",
    "seaborn~=0.12.2",
    "transformers~=4.40.2",
    "setuptools~=68.2.2",
    "wandb~=0.18.7",
    "torchaudio>=2.0.2",
    "librosa>=0.10.1",
    "numba>=0.58.0",
    "torchvision>=0.15.2",
    "segmentation-models-pytorch>=0.3.3",
    "rich>=13.6.0",
    "types-tqdm>=4.66.0.2",
    "optuna>=4.1.0",
    "jax==0.4.26",
    "jaxlib==0.4.26",
    "flax==0.8.2",
    "optax==0.2.2",
    "plotly~=5.24.1",
    "lightgbm~=4.5.0",
    "xgboost>=2.1.3",
    "catboost>=1.2.7",
    "lifelines==0.30.0",
    "category-encoders~=2.7.0",
    "torchsurv~=0.1.4",
    "rtdl_num_embeddings~=0.0.11",
    "pytorch-tabular~=1.1.1",
    # "torchsort~=0.1.9",
    "wheel>=0.45.1",

    "ninja>=1.11.1.3",

]
readme = "README.md"
requires-python = ">= 3.10"

[project.optional-dependencies]
compile = [
    "torchsort~=0.1.9",
]

[build-system]
requires = ["hatchling", "torch==2.3.0", "wheel"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/", "syamoji141/src/"]

[tool.uv.pip]
index-url = "https://download.pytorch.org/whl/cu121"

[[tool.uv.dependency-metadata]]
name = "torchsort"
version = "0.1.9"
requires-dist = ["torch"]
# [tool.uv.workspace]
# members = ["syamoji141", "kakao"]

[tool.uv]
managed = true
no-build-isolation-package = ["torchsort"]
dev-dependencies = [
    "jupyter>=1.1.1",
    "mypy==1.11.2",
    "pre-commit>=3.8.0",
    "ruff>=0.6.9",
]

[tool.mypy]
allow_redefinition = true
allow_untyped_globals = false
check_untyped_defs = true
color_output = true
disallow_incomplete_defs = true
disallow_untyped_calls = false
disallow_untyped_decorators = false
disallow_untyped_defs = true
error_summary = true
ignore_missing_imports = true
implicit_reexport = false
namespace_packages = true
no_implicit_optional = true
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
show_traceback = true
strict = true
warn_no_return = true
warn_redundant_casts = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = false

[tool.ruff.lint]
fixable = ["ALL"]
unfixable = []
select = [
    "F", # pyflakes
    "E", # pycodestyle
    "W", # pycodestyle warnings
    "I", # isort
]
ignore = [
    "E501", # line too long
]

[tool.ruff]
extend-include = ["*.ipynb"]
line-length = 99
